+++ tui/src/ui/mod.rs
@@ -174,17 +174,25 @@ impl UI {
         Ok(())
     }
 
-    /// Handle `current_track_index` possibly being changed
-    fn handle_current_track_index(&mut self, current_track_index: usize) {
+    /// Handle setting the current track index in the TUI playlist and selecting the proper list item
+    fn handle_current_track_index(&mut self, current_track_index: usize, force_relocate: bool) {
+        let tui_old_current_index = self.model.playlist.get_current_track_index();
         info!(
-            "index from player is:{current_track_index:?}, index in tui is:{:?}",
-            self.model.playlist.get_current_track_index()
+            "index from player is: {current_track_index:?}, index in tui is: {tui_old_current_index:?}"
         );
         self.model.playlist.clear_current_track();
         self.model
             .playlist
             .set_current_track_index(current_track_index);
-        self.model.playlist_locate(current_track_index);
+
+        let playlist_comp_selected_index = self.model.playlist_get_selected_index();
+
+        // only re-select the current-track if the old selection was the old-current-track
+        if force_relocate || playlist_comp_selected_index.is_none_or(|v| v == tui_old_current_index)
+        {
+            self.model.playlist_locate(current_track_index);
+        }
+
         self.model.current_song = self.model.playlist.current_track().cloned();
         self.model.update_layout_for_current_track();
         self.model.player_update_current_track_after();
@@ -248,6 +256,7 @@ impl UI {
                     if response.current_track_updated {
                         self.handle_current_track_index(
                             usize::try_from(response.current_track_index).unwrap(),
+                            false,
                         );
                     }
 
@@ -396,6 +405,7 @@ impl UI {
                     if track_changed_info.current_track_updated {
                         self.handle_current_track_index(
                             usize::try_from(track_changed_info.current_track_index).unwrap(),
+                            false,
                         );
                     }
 
@@ -443,9 +453,14 @@ impl UI {
     async fn load_playlist(&mut self) -> Result<()> {
         info!("Requesting Playlist from server");
         let tracks = self.playback.get_playlist().await?;
+        let current_track_index = tracks.current_track_index;
         self.model
             .playlist
             .load_from_grpc(tracks, &self.model.podcast.db_podcast)?;
+
+        self.model.playlist_sync();
+
+        self.handle_current_track_index(usize::try_from(current_track_index).unwrap(), true);
 
         Ok(())
     }